Fixing log mesage in test_single_register

+++ software/src/nixie_driver.cpp
@@ -86,7 +86,7 @@ void NixieDisplay::set_led(uint8_t led, bool on) {
 
 void test_single_register(NixieDisplay& display, uint8_t reg) {
 	bool curr_setting = display.get_register(reg);
-	std::cout << "\tFLIPPING REGISTER " << reg << " (CURRENTLY " << curr_setting << ")" << std::endl;
+	std::cout << "\tFLIPPING REGISTER " << (int)reg << " (CURRENTLY " << curr_setting << ")" << std::endl;
 	display.set_register(reg, !curr_setting);
 	display.update();
 }
